FIX: units must be bytes not str
